refactor: 바텀 시트 입력 필드 관련 클래스 폴더링
