+++ lib/presentation/screen/component/field/color_selection_field.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
-import '../../const/strings.dart';
-import '../../const/styles.dart';
-import '../schedule_regist/create_schedule_bottom_sheet.dart';
+import '../../../const/strings.dart';
+import '../../../const/styles.dart';
+import '../../schedule_bottom_sheet/create_schedule_bottom_sheet.dart';
 
 class ColorSelectionField extends StatelessWidget {
   final List<Color> colors;
