+++ lib/presentation/screen/schedule_bottom_sheet/create_schedule_bottom_sheet.dart
@@ -1,7 +1,7 @@
 import 'package:calendar_scheduler/domain/usecase/register_schedule.dart';
 import 'package:calendar_scheduler/presentation/const/colors.dart';
 import 'package:calendar_scheduler/presentation/extension.dart';
-import 'package:calendar_scheduler/presentation/screen/component/time_input_field.dart';
+import 'package:calendar_scheduler/presentation/screen/component/field/time_input_field.dart';
 import 'package:calendar_scheduler/presentation/util/validation_util.dart';
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
@@ -11,8 +11,8 @@ import '../../../domain/entity/schedule.dart';
 import '../../const/strings.dart';
 import '../../const/styles.dart';
 import '../provider/schedule_provider.dart';
-import '../component/color_selection_field.dart';
-import '../component/content_input_field.dart';
+import '../component/field/color_selection_field.dart';
+import '../component/field/content_input_field.dart';
 
 class CreateScheduleBottomSheet extends StatefulWidget {
   final DateTime selectedDate;
