+++ lib/presentation/screen/schedule_bottom_sheet/edit_schedule_bottom_sheet.dart
@@ -1,5 +1,5 @@
 import 'package:calendar_scheduler/presentation/extension.dart';
-import 'package:calendar_scheduler/presentation/screen/component/time_input_field.dart';
+import 'package:calendar_scheduler/presentation/screen/component/field/time_input_field.dart';
 import 'package:calendar_scheduler/presentation/util/validation_util.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
@@ -12,8 +12,8 @@ import '../../const/colors.dart';
 import '../../const/strings.dart';
 import '../../const/styles.dart';
 import '../provider/schedule_provider.dart';
-import '../component/color_selection_field.dart';
-import '../component/content_input_field.dart';
+import '../component/field/color_selection_field.dart';
+import '../component/field/content_input_field.dart';
 
 class EditScheduleBottomSheet extends StatefulWidget {
   final Schedule prevSchedule;
